build: fix the compiler warnings, including three real bugs
A clean build emitted 288 warnings; it now emits none. Most were noise, but
three were genuine defects:

  sys_public.h: xthreadInfo::threadHandle/threadId were 'unsigned long', which
  is 32 bit on Win64 (LLP64), so sys/threads.cpp stored a HANDLE in it and
  truncated the pointer. Widened to uintptr_t and the casts updated to match.

  Rotation.h: idRotation's default constructor was empty, leaving axisValid
  indeterminate - ToMat3() could therefore take the 'matrix already valid'
  path and return an uninitialised axis. Both constructors now clear
  axisValid and initialise axis, which also stops the implicit copy/move
  assignment from propagating indeterminate floats (that was the source of
  the -Wmaybe-uninitialized reports out of idPhysics_AF::Evolve).

  Common.cpp, CollisionModel_load.cpp: six '%zu'/'%zd' format strings.
  MinGW's printf has no 'z' length modifier, so these printed garbage on
  Windows. Replaced with %i and an explicit (int) cast; every value is a KB
  count or a string length.

Also fixed: MegaTexture::SetMappingForSurface() read origin/axis .xyz that
were only assigned if the vertex scan matched, so they are zeroed first;
ExtractPort() and extract_basename() replaced strncpy() (bound == destination
size, so no terminator) with an explicit length clamp, memcpy and terminator;
the save-path snprintf() is now sized so concatenation cannot truncate.

The bulk of the count was one character: '~idList<type>()' in List.h is
ill-formed in C++20 and every translation unit includes it, which alone
accounted for ~90 warnings.

Cleanups: removed dead statics (rumble, needs_gl_reinit), unused locals, and
a SIGPIPE handler that was never installed (libretro-common's net_compat.c
already does signal(SIGPIPE, SIG_IGN)); fixed a comment that opened a nested
block comment; renamed the 'requires' member/parameter to 'requiresItem'
since it is a C++20 keyword - the "requires" spawn-arg string and the
savegame read/write order are deliberately untouched, so map data and
existing savegames are unaffected. The cross-enum comparison in AF.cpp is
made explicit with casts rather than 'fixed': upstream marks it FIXME and
changing the semantics would change AF reload behaviour.

Two categories are suppressed at the build level rather than edited, because
they are id idioms rather than defects and touching ~150 sites would be churn
with real regression risk: -Wno-class-memaccess (deliberate memset/memcpy of
POD-ish structs the compiler considers non-trivial only because they have
constructors) and -Wno-write-strings (string literals stored in char* binding
tables). -Wno-unused-result is set for C so the vendored libretro-common
copies stay diffable against upstream.

+++ neo/d3xp/AF.cpp
@@ -897,8 +897,10 @@ bool idAF::Load( idEntity *ent, const char *fileName ) {
 			//  CONSTRAINT_HINGESTEERING has no DECLAF_CONSTRAINT_ equivalent,
 			//  and thus DECLAF_CONSTRAINT_SLIDER != CONSTRAINT_SLIDER (5 != 6)
 			//  and DECLAF_CONSTRAINT_SPRING != CONSTRAINT_SPRING (6 != 10)
+			// The cast makes the mismatch explicit rather than silencing a
+			// warning by accident; behaviour is unchanged from upstream.
 			if ( file->constraints[j]->name.Icmp( constraint->GetName() ) == 0 &&
-					file->constraints[j]->type == constraint->GetType() ) {
+					(int)file->constraints[j]->type == (int)constraint->GetType() ) {
 				break;
 			}
 		}
